+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+## v3.108.0
 * Added `query.EmptyTxControl()` for empty transaction control (server-side defines transaction control by internal logic)
 * Marked as deprecated `query.NoTx()` because this is wrong name for server-side transaction control inference